Created route.js file and added title img
