+++ myTrip/static/src/registration/registration.js
@@ -1,5 +1,4 @@
 import React from "react";
-import axios from "axios"
 
 import Paper from 'material-ui/Paper';
 import TextField from 'material-ui/TextField';
